requirements.txt
matplotlib
nibabel
numpy
Pillow
scipy
tqdm
pandas
monai
pytorch-lightning 
wandb
sklearn
grad-cam==1.3.5
ttach==0.0.3
pyyaml
torchmetrics
https://github.com/ufoym/imbalanced-dataset-sampler/archive/master.zip
opencv-python-headless
streamlit
st-clickable-images